fix breaking tests
