+++ stats/opentelemetry/opentelemetry.go
@@ -395,21 +395,23 @@ func DefaultMetrics() *stats.MetricSet {
 	return defaultPerCallMetrics.Join(estats.DefaultMetrics)
 }
 
-// NoopMetricsRecorder is a noop MetricsRecorder to be used in tests to prevent
+// OtelNoopMetricsRecorder is a noop MetricsRecorder to be used to prevent
 // nil panics.
-type NoopMetricsRecorder struct{}
+type OtelNoopMetricsRecorder struct{}
 
 // RecordInt64Count is a noop implementation of RecordInt64Count.
-func (r *NoopMetricsRecorder) RecordInt64Count(*estats.Int64CountHandle, int64, ...string) {}
+func (r *OtelNoopMetricsRecorder) RecordInt64Count(*estats.Int64CountHandle, int64, ...string) {}
 
 // RecordFloat64Count is a noop implementation of RecordFloat64Count.
-func (r *NoopMetricsRecorder) RecordFloat64Count(*estats.Float64CountHandle, float64, ...string) {}
+func (r *OtelNoopMetricsRecorder) RecordFloat64Count(*estats.Float64CountHandle, float64, ...string) {
+}
 
 // RecordInt64Histo is a noop implementation of RecordInt64Histo.
-func (r *NoopMetricsRecorder) RecordInt64Histo(*estats.Int64HistoHandle, int64, ...string) {}
+func (r *OtelNoopMetricsRecorder) RecordInt64Histo(*estats.Int64HistoHandle, int64, ...string) {}
 
 // RecordFloat64Histo is a noop implementation of RecordFloat64Histo.
-func (r *NoopMetricsRecorder) RecordFloat64Histo(*estats.Float64HistoHandle, float64, ...string) {}
+func (r *OtelNoopMetricsRecorder) RecordFloat64Histo(*estats.Float64HistoHandle, float64, ...string) {
+}
 
 // RecordInt64Gauge is a noop implementation of RecordInt64Gauge.
-func (r *NoopMetricsRecorder) RecordInt64Gauge(*estats.Int64GaugeHandle, int64, ...string) {}
+func (r *OtelNoopMetricsRecorder) RecordInt64Gauge(*estats.Int64GaugeHandle, int64, ...string) {}
